FPS Game version 1.00.10
I have made a FPS game which player can aim at targets and shoot them. after hit one target, the target disappears and another target replace it. there is a customized sensitivity feature for the game as well. I'm not done with this yet. However, it meet the criteria of this assignment.