Hosain and an all nighter XD, 5th April
I have coded a puzzle block game which has a 9x9 grid as a base and three randomly generated blocks to be placed in the grid. The is a dark mode for the game, and rotation option for the block. pressing R key rotates the block. I have added a Div from HTML into the js file for the Game Over pop up window. I really enjoyed doing this. I slept for 2 hours in 36 hours. the boring part commenting is not done yet. I want to make it 2D-3D if possible in the next week coming. sorry for yapping a lot, but you know the FEELIING when everything works finally.